requirements.txt
openai
requests
python-dotenv
beautifulsoup4
